Split `apply_ufunc` out of `computation.py` (#10133)
* Split `apply_ufunc` out of `computation.py`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
