+++ xarray/coding/calendar_ops.py
@@ -9,7 +9,7 @@ from xarray.coding.times import (
     _should_cftime_be_used,
     convert_times,
 )
-from xarray.computation.computation import apply_ufunc
+from xarray.computation.apply_ufunc import apply_ufunc
 from xarray.core.common import (
     _contains_datetime_like_objects,
     full_like,
