+++ xarray/computation/weighted.py
@@ -6,7 +6,8 @@ from typing import TYPE_CHECKING, Generic, Literal, cast
 import numpy as np
 from numpy.typing import ArrayLike
 
-from xarray.computation.computation import apply_ufunc, dot
+from xarray.computation.apply_ufunc import apply_ufunc
+from xarray.computation.computation import dot
 from xarray.core import duck_array_ops, utils
 from xarray.core.alignment import align, broadcast
 from xarray.core.types import Dims, T_DataArray, T_Xarray
