+++ xarray/core/missing.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING, Any, TypeVar, get_args
 import numpy as np
 import pandas as pd
 
-from xarray.computation.computation import apply_ufunc
+from xarray.computation.apply_ufunc import apply_ufunc
 from xarray.core import utils
 from xarray.core.common import _contains_datetime_like_objects, ones_like
 from xarray.core.duck_array_ops import (
